Move additional fixtures into conftest
These fixtures are all related to creating simulated PVSystem data
using pvlib.
